pyproject.toml
[project]
name = "nonebot-plugin-mcqq"
version = "2.7.0"
description = "基于NoneBot的QQ群聊与Minecraft Server消息互通插件"
authors = [{ name = "17TheWord", email = "17theword@gmail.com" }]
requires-python = ">=3.10,<4"
readme = "README.md"
license = "MIT"
dependencies = [
    "nonebot-adapter-minecraft>=1.6.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-adapter-qq>=1.6.5",
    "nonebot2[fastapi,httpx,websockets]>=2.4.3",
]

[project.optional-dependencies]
pil = [
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "nb-cli>=1.4.2",
    "pre-commit>=4.2.0,<5",
]

lint = ["ruff>=0.12.4,<0.13"]

[tool.nonebot]
adapters = [
    { name = "Minecraft", module_name = "nonebot.adapters.minecraft" },
    { name = "QQ", module_name = "nonebot.adapters.qq" },
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_mcqq"]
plugin_dirs = []
builtin_plugins = []

[tool.uv]
default-groups = ["dev", "lint"]

[tool.hatch.build.targets.sdist]
include = ["nonebot"]

[tool.hatch.build.targets.wheel]
include = ["nonebot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",
    "E501",   # module-import-not-at-top-of-file
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.isort]
force-sort-within-sections = true
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false
